changed Goods moving mechanism to reoder as whole, not per request. Add unsaved changes guard

+++ src/app/modules/admin/modules/place-list/components/place-detail/place-detail.component.ts
@@ -1,7 +1,7 @@
-import {Component, OnInit} from '@angular/core';
-import {EPlaceRole, IPlace} from '../../../../../../common/types/IPlace';
+import {Component, inject, OnInit} from '@angular/core';
+import {IPlace} from '../../../../../../common/types/IPlace';
 import {ERoute} from '../../../../../../common/types/ERoute';
-import {ActivatedRoute, Router} from '@angular/router';
+import {ActivatedRoute, NavigationStart, Router} from '@angular/router';
 import {PlaceService} from '../../../../services/place/place/place.service';
 import {CdkDragDrop, moveItemInArray} from '@angular/cdk/drag-drop';
 import {MatDialog} from '@angular/material/dialog';
@@ -9,29 +9,39 @@ import {SortimentDetailComponent} from './components/sortiment-detail/sortiment-
 import {IGoods} from '../../../../../../common/types/IGoods';
 import {AlertService} from '../../../../../../common/services/alert/alert.service';
 import {HttpErrorResponse} from '@angular/common/http';
+import {filter, takeUntil} from "rxjs";
+import {WithSubscriptionsComponent} from "../../../../../../common/components/with-subscriptions.component";
+import {CanComponentDeactivate} from "../../../../../../common/types/CanComponentDeactivate";
 
 @Component({
-	selector: 'app-place-detail',
-	templateUrl: './place-detail.component.html',
-	styleUrls: ['./place-detail.component.scss'],
+  selector: 'app-place-detail',
+  templateUrl: './place-detail.component.html',
+  styleUrls: ['./place-detail.component.scss'],
 })
-export class PlaceDetailComponent implements OnInit {
+export class PlaceDetailComponent extends WithSubscriptionsComponent implements OnInit, CanComponentDeactivate {
+  public readonly placeService = inject(PlaceService);
+  protected readonly route = inject(ActivatedRoute);
+  protected readonly router = inject(Router);
+  protected readonly dialog = inject(MatDialog);
+  protected readonly alertService = inject(AlertService);
+
 	public place: IPlace;
 	public goods: IGoods[] = [];
 	public isLoading: boolean = false;
 	public isEdit: boolean = false;
+  protected goodsPositionChanged: boolean = false;
 
-	public readonly EPlaceRole = EPlaceRole;
 	public readonly ERoute = ERoute;
 
-	constructor(
-		public placeService: PlaceService,
-		protected route: ActivatedRoute,
-		protected router: Router,
-		protected dialog: MatDialog,
-		protected alertService: AlertService,
-	) {
-	}
+  public canDeactivate(): boolean {
+    if (!this.goodsPositionChanged) {
+      return true;
+    }
+
+    return window.confirm(
+      'Pozice zboží nebyla uložena, opravdu chcete odejít?'
+    );
+  };
 
 	public async ngOnInit(): Promise<void> {
 		this.isLoading = true;
@@ -51,6 +61,13 @@ export class PlaceDetailComponent implements OnInit {
 		} finally {
 			this.isLoading = false;
 		}
+
+    this.router.events
+      .pipe(
+        takeUntil(this.destroy$),
+        filter((e) => e instanceof NavigationStart)
+      )
+      .subscribe()
 	}
 
 
@@ -76,7 +93,7 @@ export class PlaceDetailComponent implements OnInit {
 
 	public async drop(event: CdkDragDrop<string[]>): Promise<void> {
 		try {
-			await this.placeService.moveGoods(this.place!.id!, this.goods[event.previousIndex].id!, this.goods[event.currentIndex].id!);
+      this.goodsPositionChanged = true;
 			moveItemInArray(this.goods, event.previousIndex, event.currentIndex);
 		} catch(e) {
 			console.error('Canot move goods: ', e);
@@ -84,12 +101,25 @@ export class PlaceDetailComponent implements OnInit {
 		}
 	}
 
+  protected async confirmPosition(): Promise<void> {
+    try {
+      await this.placeService.moveGoods(this.place!.id!, this.goods);
+      this.goodsPositionChanged = false;
+      this.alertService.success('Pozice zboží změněna');
+    } catch(e) {
+      console.error('Canot move goods: ', e);
+      this.alertService.error('Nepovedlo se změnit pořadí');
+    }
+  }
+
 	public async removeItem(id: number): Promise<void> {
 		try {
-			await this.placeService.removeGoods(id, this.place!.id!);
+      if(this.isEdit) {
+        await this.placeService.removeGoods(id, this.place!.id!);
+      }
 			this.goods = this.goods.filter((obj) => obj.id !== id);
 		} catch(e) {
-			this.alertService.success('Nepodařilo se odebrat zboží');
+			this.alertService.error('Nepodařilo se odebrat zboží');
 			console.error('Cannot remove item', e)
 		}
 	}
@@ -117,8 +147,8 @@ export class PlaceDetailComponent implements OnInit {
 					for(const item of result) {
 						await this.placeService.addGoods(item.id, this.place.id);
 					}
-					this.goods.push(...result);
 				}
+        this.goods.push(...result);
 			} catch(e) {
 				if(e instanceof HttpErrorResponse) {
 					this.alertService.error(e.error.Message ?? 'Chyba při přidávání sortimentu');
